Organize UI tests.
