--- Example/PrimerSDKExample_UITests/ApplePay.swift
@@ -1,64 +1,0 @@
-//
-//  ApplePay.swift
-//  PrimerSDKExample_UITests
-//
-//  Created by Evangelos on 3/12/21.
-//  Copyright © 2021 CocoaPods. All rights reserved.
-//
-
-import XCTest
-
-class ApplePay: XCTestCase {
-    
-    let app = XCUIApplication()
-    let base = Base()
-
-    override func setUp() {
-        super.setUp()
-        app.launchArguments = ["UITesting"]
-        app.launch()
-    }
-
-    override func tearDown() {
-        super.tearDown()
-    }
-    
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-        XCUIApplication().launch()
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-    
-    func testPresentApplePay() throws {
-        let pm = Base.paymentMethods.filter({ $0.id == "APPLE_PAY" }).first!
-        
-        try base.testInitialize(
-            env: pm.environment.rawValue,
-            customerId: nil,
-            phoneNumber: nil,
-            countryCode: pm.countryCode,
-            currency: pm.currency,
-            amount: pm.amount,
-            performPayment: true)
-
-        try base.openUniversalCheckout()
-
-        if let amountExpectation = pm.expecations?.amount {
-            let amountText = app.staticTexts[amountExpectation]
-            XCTAssert(amountText.exists, "Amount '\(amountExpectation)' should exist")
-        }
-        
-        let scrollView = app.scrollViews["primer_container_scroll_view"]
-        if let surchargeExpectation = pm.expecations?.surcharge {
-            Base.validateSurcharge(surchargeExpectation, forPaymentMethod: pm.id)
-        }
-        
-        let applePayButton = scrollView.buttons[pm.id]
-        applePayButton.tap()
-    }
-
-}
-
