+++ Example/PrimerSDKExample_UITests/ExternalPaymentMethods.swift
@@ -31,201 +31,7 @@ class ExternalPaymentMethods: XCTestCase {
     override func tearDownWithError() throws {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
-    
-    func testAdyenAlipay() throws {
-        let pm = Base.paymentMethods.filter({ $0.id == "ADYEN_ALIPAY" }).first!
-        try testPayment(pm)
-    }
-    
-    func testAdyenGiropay() throws {
-        let payment = Base.paymentMethods.filter({ $0.id == "ADYEN_GIROPAY" }).first!
-        try testPayment(payment, cancelPayment: false)
-        
-        let safariWebView = app.otherElements.matching(NSPredicate(format: "identifier CONTAINS 'BrowserView?WebViewProcessID'"))
-        
-        let continueButton = app.webViews.firstMatch.buttons.firstMatch
-        let continueButtonIsHittable = expectation(for: Expectation.isHittable, evaluatedWith: continueButton, handler: nil)
-        wait(for: [continueButtonIsHittable], timeout: 30)
-        continueButton.tap()
-        
-        let scTextField = app.otherElements.containing(NSPredicate(format: "label == 'sc:'")).firstMatch.textFields.firstMatch
-        let scTextFieldIsHittable = expectation(for: Expectation.isHittable, evaluatedWith: scTextField, handler: nil)
-        wait(for: [scTextFieldIsHittable], timeout: 30)
-        scTextField.tap()
-        scTextField.typeText("10")
-        
-        let toolBarDoneButton = app.toolbars["Toolbar"].firstMatch.buttons["Done"].firstMatch
-        toolBarDoneButton.tap()
-        
-        let extensionScTextField = app.otherElements.containing(NSPredicate(format: "label == 'extensionSc:'")).firstMatch.textFields.element(boundBy: 1)
-        extensionScTextField.tap()
-        extensionScTextField.typeText("4000")
-        
-        toolBarDoneButton.tap()
-        safariWebView.webViews.firstMatch.pinch(withScale: 0.5, velocity: -0.5)
-        
-        let customerNameTextField = app.otherElements.containing(NSPredicate(format: "label == 'customerName1:'")).firstMatch.textFields.element(boundBy: 2)
-        customerNameTextField.tap()
-        customerNameTextField.typeText("John Smith")
-        
-        toolBarDoneButton.tap()
-        safariWebView.webViews.firstMatch.pinch(withScale: 0.5, velocity: -0.5)
-        
-        let customerIbanTextField = app.otherElements.containing(NSPredicate(format: "label == 'customerIBAN:'")).firstMatch.textFields.element(boundBy: 3)
-        customerIbanTextField.tap()
-        customerIbanTextField.typeText("DE36444488881234567890")
-        
-        toolBarDoneButton.tap()
-        safariWebView.webViews.firstMatch.pinch(withScale: 0.5, velocity: -0.5)
-        
-        let submitButton = app.buttons.matching(NSPredicate(format: "label == 'Absenden'")).firstMatch
-        submitButton.tap()
-        
-        try base.testSuccessMessageExists()
-        try base.testDismissSDK()
-        try base.testResultScreenExpectations(for: payment)
-    }
-    
-    func testAdyenMobilePay() throws {
-        let pm = Base.paymentMethods.filter({ $0.id == "ADYEN_MOBILEPAY" }).first!
-        try testPayment(pm)
-    }
-    
-    func testPayNLBancontact() throws {
-        let pm = Base.paymentMethods.filter({ $0.id == "PAY_NL_BANCONTACT" }).first!
-        try testPayment(pm)
-    }
-    
-    func testPresentApplePay() throws {
-        let pm = Base.paymentMethods.filter({ $0.id == "APPLE_PAY" }).first!
-        
-        try base.testInitialize(
-            env: pm.environment.rawValue,
-            customerId: nil,
-            phoneNumber: nil,
-            countryCode: pm.countryCode,
-            currency: pm.currency,
-            amount: pm.amount,
-            performPayment: true)
 
-        try base.openUniversalCheckout()
-
-        if let amountExpectation = pm.expecations?.amount {
-            let amountText = app.staticTexts[amountExpectation]
-            XCTAssert(amountText.exists, "Amount '\(amountExpectation)' should exist")
-        }
-        
-        let scrollView = app.scrollViews["primer_container_scroll_view"]
-        if let surchargeExpectation = pm.expecations?.surcharge {
-            Base.validateSurcharge(surchargeExpectation, forPaymentMethod: pm.id)
-        }
-        
-        let applePayButton = scrollView.buttons[pm.id]
-        applePayButton.tap()
-    }
-    
-    func testPayment(_ payment: Payment, cancelPayment: Bool = true) throws {
-        try Base().testInitialize(
-            env: payment.environment.rawValue,
-            customerId: nil,
-            phoneNumber: nil,
-            countryCode: payment.countryCode,
-            currency: payment.currency,
-            amount: payment.amount,
-            performPayment: true)
-
-        let universalCheckoutButton = app.buttons["universal_checkout_button"]
-        universalCheckoutButton.tap()
-
-        // Test that title is correct
-        let checkoutTitle = app.staticTexts["Choose payment method"]
-        let vaultTitle = app.staticTexts["Add payment method"]
-
-        
-        expectation(for: Expectation.exists, evaluatedWith: checkoutTitle, handler: nil)
-        expectation(for: Expectation.doesNotExist, evaluatedWith: vaultTitle, handler: nil)
-        waitForExpectations(timeout: 30, handler: nil)
-
-        if let amountExpectation = payment.expecations?.amount {
-            let amountText = app.staticTexts[amountExpectation]
-            XCTAssert(amountText.exists, "Amount '\(amountExpectation)' should exist")
-        }
-        
-        let scrollView = app.scrollViews["primer_container_scroll_view"]
-        if let surchargeExpectation = payment.expecations?.surcharge {
-            Base.validateSurcharge(surchargeExpectation, forPaymentMethod: payment.id)
-        }
-        
-        let paymentMethodButton = app.buttons[payment.id]
-        
-        if !paymentMethodButton.exists {
-            var isHittable: Bool = false
-            while !isHittable {
-                scrollView.swipeUp()
-                isHittable = paymentMethodButton.isHittable
-            }
-        }
-        
-        let adyenGiropayButton = scrollView.otherElements.buttons[payment.id]
-        adyenGiropayButton.tap()
-        
-        let webViews = app.webViews
-        if let webViewImageExpectation = payment.expecations?.webviewImage {
-            let webViewGiroPayImage = webViews.images[webViewImageExpectation]
-            let webViewGiroPayImageExists = expectation(for: Expectation.exists, evaluatedWith: webViewGiroPayImage, handler: nil)
-            wait(for: [webViewGiroPayImageExists], timeout: 30)
-        }
-        
-        if let webviewTexts = payment.expecations?.webviewTexts {
-            var webviewTextsExpectations: [XCTestExpectation] = []
-            for text in webviewTexts {
-                let webViewText = webViews.staticTexts[text]
-                let webViewTextExists = expectation(for: Expectation.exists, evaluatedWith: webViewText, handler: nil)
-                webviewTextsExpectations.append(webViewTextExists)
-                
-            }
-            
-            wait(for: webviewTextsExpectations, timeout: 30)
-        }
-        
-        if cancelPayment {
-            let safariDoneButton = app.otherElements["TopBrowserBar"].buttons["Done"]
-            safariDoneButton.tap()
-            let canceledLabel = app.scrollViews["primer_container_scroll_view"].otherElements.staticTexts["User cancelled"]
-            let canceledLabelExists = expectation(for: Expectation.exists, evaluatedWith: canceledLabel, handler: nil)
-            wait(for: [canceledLabelExists], timeout: 3)
-            
-            scrollView.swipeDown()
-            
-            if let resultScreenTextExpectations = payment.expecations?.resultScreenTexts {
-                var expectations: [XCTestExpectation] = []
-                
-                if let status = resultScreenTextExpectations["status"] as? String {
-                    let statusText = app.staticTexts[status]
-                    let statusTextExists = expectation(for: Expectation.exists, evaluatedWith: statusText, handler: nil)
-                    expectations.append(statusTextExists)
-                }
-                
-                if let actions = resultScreenTextExpectations["actions"] as? String {
-                    let actionsText = app.staticTexts[actions]
-                    let actionsTextExists = expectation(for: Expectation.exists, evaluatedWith: actionsText, handler: nil)
-                    expectations.append(actionsTextExists)
-                }
-                
-                if let amount = resultScreenTextExpectations["amount"] as? String {
-                    let amountText = app.staticTexts[amount]
-                    let amountTextExists = expectation(for: Expectation.exists, evaluatedWith: amountText, handler: nil)
-                    expectations.append(amountTextExists)
-                }
-                
-                if !expectations.isEmpty {
-                    wait(for: expectations, timeout: 3)
-                }
-            }
-        }
-    }
-    
-    
 //    func wait(forWebViewElement element: XCUIElementTypeQueryProvider, timeout: TimeInterval = 20) {
 //        // xcode has bug, so we cannot directly access webViews XCUIElements
 //        // as a workaround we can check debugDesciption and parse it, that works
